Mark deps from Github on install & playground up
For `install_requires` of the form of `docassemble.ALToolbox @ https://github.com/SuffolkLITLab/docassemble-ALToolbox/archive/main.zip`,
will correctly note that the dependency is from Github, not pypi. Additionally
does part of that work for uploading a package to the playground (untested).

Still needs to correctly write out the dependencies when downloading a package
from the playground, i.e. in `files.py:make_package_dir`, and
`update.py:get_pip_info`. Still an open question; how do you determine where a
dependency is from? What if the original setup.py was installed from Github on
a branch, but there's a different branch on the server, or it's from pypi?